Time: 6 ms (32.58%), Space: 79.5 MB (12.22%) - LeetHub

+++ 2017-grid-game/2017-grid-game.cpp
@@ -8,9 +8,8 @@ public:
         presum[1][0] = grid[1][0];
         presum[0][n - 1] = grid[0][n - 1];
         for (int i = 1; i < n; i++) {
-            presum[1][i] = (long long)grid[1][i] + presum[1][i - 1];
-            presum[0][n - i - 1] =
-                (long long)grid[0][n - i - 1] + presum[0][n - i];
+            presum[1][i] = grid[1][i] + presum[1][i - 1];
+            presum[0][n - i - 1] = grid[0][n - i - 1] + presum[0][n - i];
         }
         long long p;
         long long ans = min(presum[0][1], presum[1][n - 2]);
